added support for different implementations
