+++ sparql/src/main/java/org/apache/zeppelin/sparql/SparqlInterpreter.java
@@ -19,22 +19,9 @@ package org.apache.zeppelin.sparql;
 
 import org.apache.commons.lang3.StringUtils;
 
-import org.apache.http.HttpStatus;
-
-import org.apache.jena.query.QueryExecution;
-import org.apache.jena.query.QueryExecutionFactory;
-import org.apache.jena.query.QueryParseException;
-import org.apache.jena.query.ResultSet;
-import org.apache.jena.query.ResultSetFormatter;
-import org.apache.jena.shared.PrefixMapping;
-import org.apache.jena.sparql.ARQException;
-import org.apache.jena.sparql.engine.http.QueryExceptionHTTP;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayOutputStream;
-import java.util.Map;
 import java.util.Properties;
 
 import org.apache.zeppelin.interpreter.Interpreter;
@@ -48,15 +35,26 @@ import org.apache.zeppelin.interpreter.InterpreterResult;
 public class SparqlInterpreter extends Interpreter {
   private static final Logger LOGGER = LoggerFactory.getLogger(SparqlInterpreter.class);
 
+  public static final String SPARQL_ENGINE_TYPE = "sparql.engine";
   public static final String SPARQL_SERVICE_ENDPOINT = "sparql.endpoint";
   public static final String SPARQL_REPLACE_URIS = "sparql.replaceURIs";
   public static final String SPARQL_REMOVE_DATATYPES = "sparql.removeDatatypes";
 
-  private String serviceEndpoint;
-  private boolean replaceURIs;
-  private boolean removeDatatypes;
+  public static final String ENGINE_TYPE_JENA = "jena";
 
-  private QueryExecution queryExecution;
+  public SparqlEngine engine;
+  
+  /**
+   * Sparql Engine Type.
+   */
+  public enum SparqlEngineType {
+    JENA {
+      @Override
+      public String toString() {
+        return ENGINE_TYPE_JENA;
+      }
+    }
+  }
 
   public SparqlInterpreter(Properties properties) {
     super(properties);
@@ -66,90 +64,35 @@ public class SparqlInterpreter extends Interpreter {
   public void open() {
     LOGGER.info("Properties: {}", getProperties());
 
-    serviceEndpoint = getProperty(SPARQL_SERVICE_ENDPOINT);
-    replaceURIs = getProperty(SPARQL_REPLACE_URIS) != null
+    String serviceEndpoint = getProperty(SPARQL_SERVICE_ENDPOINT);
+    boolean replaceURIs = getProperty(SPARQL_REPLACE_URIS) != null
         && getProperty(SPARQL_REPLACE_URIS).equals("true");
-    removeDatatypes = getProperty(SPARQL_REMOVE_DATATYPES) != null
+    boolean removeDatatypes = getProperty(SPARQL_REMOVE_DATATYPES) != null
         && getProperty(SPARQL_REMOVE_DATATYPES).equals("true");
+    String engineType = getProperty(SPARQL_ENGINE_TYPE);
 
-    queryExecution = null;
+    if (SparqlEngineType.JENA.toString().equals(engineType)) {
+      engine = new JenaInterpreter(serviceEndpoint, replaceURIs, removeDatatypes);
+    } 
   }
 
   @Override
   public void close() {
-    if (queryExecution != null) {
-      queryExecution.close();
-    }
+    engine.close();
   }
 
   @Override
   public InterpreterResult interpret(String query, InterpreterContext context) {
-    LOGGER.info("SPARQL: Run Query '" + query + "' against " + serviceEndpoint);
-
     if (StringUtils.isEmpty(query) || StringUtils.isEmpty(query.trim())) {
       return new InterpreterResult(InterpreterResult.Code.SUCCESS);
     }
 
-    try {
-      queryExecution = QueryExecutionFactory.sparqlService(serviceEndpoint, query);
-      PrefixMapping prefixMapping = queryExecution.getQuery().getPrefixMapping();
-
-      // execute query and get Results
-      ResultSet results = queryExecution.execSelect();
-
-      // transform ResultSet to TSV-String
-      ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-      ResultSetFormatter.outputAsTSV(outputStream, results);
-      String tsv = new String(outputStream.toByteArray());
-
-      if (replaceURIs) {
-        LOGGER.info("SPARQL: Replacing URIs");
-        tsv = replaceURIs(tsv, prefixMapping);
-      }
-
-      if (removeDatatypes) {
-        LOGGER.info("SPARQL: Removing datatypes");
-        tsv = removeDatatypes(tsv);
-      }
-
-      return new InterpreterResult(
-              InterpreterResult.Code.SUCCESS,
-              InterpreterResult.Type.TABLE,
-              tsv);
-    } catch (QueryParseException e) {
-      LOGGER.error(e.toString());
-      return new InterpreterResult(
-        InterpreterResult.Code.ERROR,
-        "Error: " + e.getMessage());
-    } catch (QueryExceptionHTTP e) {
-      LOGGER.error(e.toString());
-      int responseCode = e.getResponseCode();
-
-      if (responseCode == HttpStatus.SC_UNAUTHORIZED) {
-        return new InterpreterResult(
-          InterpreterResult.Code.ERROR,
-            "Unauthorized.");
-      } else if (responseCode == HttpStatus.SC_NOT_FOUND) {
-        return new InterpreterResult(
-          InterpreterResult.Code.ERROR,
-            "Endpoint not found, please check endpoint in the configuration.");
-      } else {
-        return new InterpreterResult(
-          InterpreterResult.Code.ERROR,
-            "Error: " + e.getMessage());
-      }
-    } catch (ARQException e) {
-      return new InterpreterResult(
-        InterpreterResult.Code.INCOMPLETE,
-          "Query cancelled.");
-    }
+    return engine.query(query);
   }
 
   @Override
   public void cancel(InterpreterContext context) {
-    if (queryExecution != null) {
-      queryExecution.abort();
-    }
+    engine.cancel();
   }
 
   @Override
@@ -160,18 +103,5 @@ public class SparqlInterpreter extends Interpreter {
   @Override
   public int getProgress(InterpreterContext context) {
     return 0;
-  }
-
-  private String replaceURIs(String tsv, PrefixMapping prefixMapping) {
-    Map<String, String> pmap = prefixMapping.getNsPrefixMap();
-    for (Map.Entry<String, String> entry : pmap.entrySet()) {
-      tsv = tsv.replaceAll(entry.getValue(), entry.getKey() + ":");
-    }
-    return tsv;
-  }
-
-  private String removeDatatypes(String tsv) {
-    // capture group: "($1)"^^<.+?>
-    return tsv.replaceAll("\"(.+?)\"\\^\\^\\<.+?\\>", "$1");
   }
 }
